% Generated by roxygen2 (4.0.2): do not edit by hand
\name{climStack2List}
\alias{climStack2List}
\title{"climStack2List" coercion of a RasterStack to a list}
\usage{
climStack2List(r)
}
\arguments{
\item{r}{\code{RasterStack-class} or a \code{RasterBrick-class} object}
}
\description{
This function transforms a \code{RasterStack-class} or a \code{RasterBrick-class} object into a list of XY elements
}
\examples{
library(ClimClassMap)
wpath <- system.file("raster_maps/1981-2010",package="ClimClassMap")
prec_map  <- "p/p1981-2010\%02d.tif"


fact <- 5

prec <- aggregate(stack(paste(wpath,sprintf(prec_map,1:12),sep="/")),fact=fact) ## 1000 m resolution
names(prec) <- sprintf("M\%02d",1:12)

l <- climStack2List(prec)
}
\seealso{
\code{\link{resterClimApply}},\code{\link{listClimApply}}
}

