% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climApply.R
\name{climApply}
\alias{climApply}
\title{climApply}
\usage{
climApply(P = NA, Tx = NA, Tm = NA, Tn = NA, month = 1:12,
  clim_fun = ClimClass::contin, ...)
}
\arguments{
\item{P}{monthly precipitation}

\item{Tx}{monthly mean daily maximium temparature}

\item{Tm}{monthly mean daily average temparature}

\item{Tn}{monthly mean daily minimum temparature}

\item{month}{vector indicating each month. Default is \code{1:12}.}

\item{clim_fun}{function or function name of the applied "ClimClass" package function}

\item{...}{further arguments for \code{clim_fun}}
}
\description{
Extantsions of 'climClass' function
}
\note{
'month' can be replecad by 'layer'
}
\examples{

library(ClimClassMap)

data(Trent_climate)
 
clim <- clima_81_10$T0083

c <- climApply(P=clim$P,Tx=clim$Tx,Tn=clim$Tn,Tm=clim$Tm,clim_fun=contin,indices=4)



}

