% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listClimApply}
\alias{listClimApply}
\title{Extension of  "ClimApply" to a list}
\usage{
listClimApply(P = NA, Tx = NA, Tm = NA, Tn = NA, month = list(1:12),
  clim_fun = "contin", ...)
}
\description{
Function "ClimApply" to ...
}
\note{
'month' can be replecad by 'layer'
}
\examples{
library(ClimClassMap)

data(Trent_climate)
clim <- melt(clima_81_10,id="month")

names(clim)[names(clim)=="L1"] <- "station"

station <- unique(clim$station)[4:6]
names(station) <- station
P <- lapply(X=station,FUN=function(it,df,variable) {df$value[df$station==it & df$variable==variable]},df=clim,variable="P")
Tx <- lapply(X=station,FUN=function(it,df,variable) {df$value[df$station==it & df$variable==variable]},df=clim,variable="Tx")
Tn <- lapply(X=station,FUN=function(it,df,variable) {df$value[df$station==it & df$variable==variable]},df=clim,variable="Tn")
Tm <- lapply(X=station,FUN=function(it,df,variable) {df$value[df$station==it & df$variable==variable]},df=clim,variable="Tm")

c <- listClimApply(P=P,Tx=Tx,Tn=Tn,Tm=Tm,clim_fun="contin",indices=4)
}
\seealso{
\code{\link{ClimApply}}
}

