% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listClimApply}
\alias{listClimApply}
\title{Extension of  "ClimApply" to a list}
\usage{
listClimApply(P = NA, Tx = NA, Tm = NA, Tn = NA, month = list(1:12),
  clim_fun = "contin", ...)
}
\arguments{
\item{P}{monthly precipitation (list)}

\item{Tx}{monthly mean daily maximium temparature (list)}

\item{Tm}{monthly mean daily average temparature (list)}

\item{Tn}{monthly mean daily minimum temparature (list)}

\item{month}{a list with a vector or vectors indicating each month. Default is \code{list(1:12)}.}

\item{clim_fun}{character string containing the name of the applied "ClimClass" package function}

\item{...}{further arguments for \code{clim_fun} function}
}
\description{
Function "ClimApply" applied to a list
}
\note{
'month' can be replecad by 'layer'
}
\examples{
library(ClimClassMap)

data(Trent_climate)
clim <- melt(clima_81_10,id="month")

names(clim)[names(clim)=="L1"] <- "station"

station <- unique(clim$station)[4:6]
names(station) <- station
P <- lapply(X=station,FUN=function(it,df,variable) {df$value[df$station==it & df$variable==variable]},df=clim,variable="P")
Tx <- lapply(X=station,FUN=function(it,df,variable) {df$value[df$station==it & df$variable==variable]},df=clim,variable="Tx")
Tn <- lapply(X=station,FUN=function(it,df,variable) {df$value[df$station==it & df$variable==variable]},df=clim,variable="Tn")
Tm <- lapply(X=station,FUN=function(it,df,variable) {df$value[df$station==it & df$variable==variable]},df=clim,variable="Tm")

c <- listClimApply(P=P,Tx=Tx,Tn=Tn,Tm=Tm,clim_fun="contin",indices=4)
}
\seealso{
\code{\link{climApply}}
}

