% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterClimApply.R
\name{rasterClimApply}
\alias{rasterClimApply}
\title{Extension of  "ClimApply" to a Raster}
\usage{
rasterClimApply(P = NULL, Tx = NULL, Tm = NULL, Tn = NA, month = NA,
  clim_fun = "contin", ...)
}
\arguments{
\item{P, Tx, Tm, Tn}{daily precipitation, maximum, mean and minimum temperature.}

\item{month}{vectors of months. Default is \code{NA}}

\item{clim_fun}{climate function index of \code{ClimClass} package}

\item{...}{further arguments for \code{\link{listClimApply}}}
}
\description{
Function "ClimApply" to ...
}
\examples{

library(ClimClassMap)
wpath <- system.file("raster_maps/1981-2010",package="ClimClassMap")
temp_map  <- "t/t1981-2010\%02d.tif" 
prec_map  <- "p/p1981-2010\%02d.tif"


fact <- 5

prec <- aggregate(stack(paste(wpath,
			sprintf(prec_map,1:12),sep="/")),fact=fact) ## 1000 m resolution 
Tm <- aggregate(stack(paste(wpath,sprintf(temp_map,1:12),sep="/")),fact=fact)

names(prec) <- sprintf("M\%02d",1:12)
names(Tm) <- sprintf("M\%02d",1:12)

# Actally no info for daily maximum and minimum temeparature

Tn <- stack(Tm-5)
Tx <- stack(Tm+5)




indexMap <- rasterClimApply(P=prec,Tx=Tx,Tn=Tn,Tm=Tm,clim_fun="contin",indices=4)




}
\seealso{
\code{\link{climApply}},\code{\link{listClimApply}}

\code{\link{climStack2List}},\code{\link{listClimApply}}
}

