% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rasterClimApply}
\alias{rasterClimApply}
\title{Extension of  "ClimApply" to a Raster}
\usage{
rasterClimApply(P = NA, Tx = NA, Tm = NA, Tn = NA, month = list(1:12),
  clim_fun = "contin", ...)
}
\description{
Function "ClimApply" to ...
}
\examples{
TO DO
}
\seealso{
\code{\link{ClimApply}},\code{\link{listClimApply}}
}

